+++ gendarme/FSharpExamples/AvoidLongParameterListsRule .fs
@@ -1,0 +1,9 @@
+﻿namespace AvoidLongParameterLists
+
+type SeqPnt =
+  { StartLine : int
+    StartColumn : int
+    EndLine : int
+    EndColumn : int
+    Document : string
+    Offset : int }