+++ ReleaseNotes.md
@@ -6,6 +6,8 @@ F# niggles part 3
 * Exempt match on union types from `AvoidSwitchStatementsRule`
 * Exempt F# generated types with `@` in the name from `UseCorrectPrefixRule`,`VariableNamesShouldNotMatchFieldNamesRule` and `UseCorrectCasingRule`
 * Exempt generated abstract closure types from `AbstractTypesShouldNotHavePublicConstructorsRule`
+* Exempt constructors of record types, or generated types with `@` in the name, from `AvoidLongParameterListsRule`
+* Module-bound functions should be camel-cased in the `UseCorrectCasingRule`
 
 For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/master/ReleaseNotes%20-%20Previously.md
 Since this is a release from a fork, [issues should be reported at my related repo](https://github.com/SteveGilham/altcode.fake/issues) that contains a Fake driver for the Gendarme tool, but noted as being against the forked Gendarme tool itself.